# Speed up complex.pow when the exponent is 2.0 or 0.5